+++ src/main/java/entities/SaveFile.java
@@ -5,16 +5,16 @@ package entities;
  */
 
 public class SaveFile {
-    private final double timeLeft;
+    private final int timeLeft;
     private final String name;
     private final String spritePath;
-    private final double hunger;
-    private final double thirst;
-    private final double cleanliness;
-    private final double happiness;
+    private final int hunger;
+    private final int thirst;
+    private final int cleanliness;
+    private final int happiness;
 
-    public SaveFile(double timeLeft, String name, String spritePath, double hunger, double thirst, double cleanliness,
-                    double happiness) {
+    public SaveFile(int timeLeft, String name, String spritePath, int hunger, int thirst, int cleanliness,
+                    int happiness) {
         this.timeLeft = timeLeft;
         this.name = name;
         this.spritePath = spritePath;
@@ -24,7 +24,7 @@ public class SaveFile {
         this.happiness = happiness;
     }
 
-    public double getTimeLeft() {
+    public int getTimeLeft() {
         return timeLeft;
     }
 
@@ -36,17 +36,17 @@ public class SaveFile {
         return spritePath;
     }
 
-    public double getHunger() {
+    public int getHunger() {
         return hunger;
     }
 
-    public double getThirst() {
+    public int getThirst() {
         return thirst;
     }
 
-    public double getCleanliness() {
+    public int getCleanliness() {
         return cleanliness;
     }
 
-    public double getHappiness() { return happiness; }
+    public int getHappiness() { return happiness; }
 }
